+++ dd-java-agent/instrumentation/javax-naming/build.gradle
