+++ dd-java-agent/instrumentation/javax-naming/src/test/groovy/InitialDirContextCallSiteTest.groovy
