+++ settings.gradle
@@ -201,7 +201,7 @@ include ':dd-java-agent:instrumentation:java-io'
 include ':dd-java-agent:instrumentation:java-lang'
 include ':dd-java-agent:instrumentation:java-lang:java-lang-9'
 include ':dd-java-agent:instrumentation:java-security'
-include ':dd-java-agent:instrumentation:javax-naming-directory'
+include ':dd-java-agent:instrumentation:javax-naming'
 include ':dd-java-agent:instrumentation:jax-rs-annotations-1'
 include ':dd-java-agent:instrumentation:jax-rs-annotations-2'
 include ':dd-java-agent:instrumentation:jax-rs-annotations-2:filter-jersey'
